refactor: BaseButtonHelper 클래스를 ButtonHelper로 통합 및 불필요한 파일 삭제
- BaseButtonHelper 클래스를 ButtonHelper로 통합하여 코드 중복 제거
- BaseButtonHelper.ts 파일 삭제로 프로젝트 구조 간소화
