+++ frontend/__tests__/playwright/helpers/shared/index.ts
@@ -1,4 +1,3 @@
-export { BaseButtonHelper } from './BaseButtonHelper';
 export { ButtonHelper } from './ButtonHelper';
 export { FormHelper } from './FormHelper';
 export { NavigationHelper } from './NavigationHelper';
